Fixes for templated function results
As part of instantiating the class templates, a new Typemap was
created.  The class Typemap will now keep track of instantiated
Typemaps via Typemap.cxx_instantiation.  Added a pass over the libary
which replaces the original TypeMap with the templated Typemap.  This
will create the proper C++ code for C_memory_dtor_function.

Added a test to templates.yaml similar to:

    - decl: template<typename T> class A
      cxx_template:
      - instantiation: <int>
    - decl: A<int> func2()

+++ regression/run/templates/templates.hpp
@@ -56,7 +56,7 @@ public:
   { }
 };
 
-user<double> returnUserType(void);
+user<int> returnUserType(void);
 
 template<typename T>
 struct structAsClass {
